Reduce filesize of some images
I saw the build size is very close to being over the 250mb limit.
The correct thing to do would be to move all static images to blob
storage and not save it in the repo. But it order to eek out maybe
another year or two before I need to do that, I can just reduce the
quality of some of the largest images.

Done.